server/requirements.txt
openenv-core>=0.2.0
openai>=1.0.0
pydantic>=2.5.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0